backend/requirements.txt
# Core Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
starlette>=0.27.0

# Database
motor>=3.0.0
pymongo>=4.0.0

# Environment & Config
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0
email-validator>=2.0.0

# AI & LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langsmith>=0.1.0

# Google AI
google-generativeai>=0.3.0
google-ai-generativelanguage>=0.4.0

# HTTP Clients
httpx>=0.24.0
requests>=2.28.0
aiohttp>=3.8.0

# Utilities
python-jose>=3.3.0
passlib>=1.7.0
bcrypt>=4.0.0
PyJWT>=2.6.0
